# api/openapi.yaml
swagger: "2.0"
info:
  title: Configuration Management
  version: 0.0.1
  description: |
    REST API for managing system resources with idempotent semantics. Includes support for
    optimistic concurrency using ETags.
consumes:
  - "application/json"
produces:
  - "application/json"
schemes:
  - http
  - https
basePath: "/api/v1/"
...
paths:
...
  /files:
    get:
      summary: Retrieve the current state and properties of a file
      description: |
        Fetches properties about the specified file, including its permissions, ownership
        and hash checksum.
      operationId: getFileProperties
      tags:
        - Files
      parameters:
        - $ref: "#/parameters/FilePath"
      responses:
        200:
          $ref: "#/responses/FilePropertiesResponse"
        400:
          description: Invalid request or missing fields
          schema:
            $ref: "#/responses/ErrorResponse"
        404:
          description: File not found
          schema:
            $ref: "#/responses/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/responses/ErrorResponse"
    put:
      summary: Create or update properties for a file
      description: |
        Creates a new file if it does not exist or updates properties (mode, owner, group)
        if the file exists. Operation is idempotent: repeating the same request yields the
        same result. File content must be managed separately.
      operationId: putFile
      tags:
        - Files
      parameters:
        - $ref: "#/parameters/FilePath"
        - $ref: "#/parameters/IfMatch"
        - in: body
          name: properties
          required: true
          schema:
            $ref: "#/definitions/FileProperties"
      responses:
        201:
          description: File created and properties were applied
          headers:
            ETag:
              type: string
              description: New ETag for the updated file
        204:
          description: File already existed and properties were either applied or already correct
          headers:
            ETag:
              type: string
              description: New ETag for the updated file
        400:
          description: Invalid request, bad path or unresolvable owner/group
          schema:
            $ref: "#/responses/ErrorResponse"
        409:
          description: Conflict due to conditional check failure (e.g. ETag mismatch)
          schema:
            $ref: "#/responses/ErrorResponse"
        412:
          description: Precondition failed
          schema:
            $ref: "#/responses/ErrorResponse"
        428:
          description: Missing If-Match header
          schema:
            $ref: "#/responses/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/responses/ErrorResponse"
    delete:
      summary: Delete an existing file
      description: Removes the specified file from the system
      operationId: deleteFile
      tags:
        - Files
      parameters:
        - $ref: "#/parameters/FilePath"
        - $ref: "#/parameters/IfMatch"
      responses:
        204:
          description: File deleted
        400:
          description: Invalid request or missing fields
          schema:
            $ref: "#/responses/ErrorResponse"
        403:
          description: Forbidden
          schema:
            $ref: "#/responses/ErrorResponse"
        409:
          description: Conflict due to conditional check failure (e.g. ETag mismatch)
          schema:
            $ref: "#/responses/ErrorResponse"
        428:
          description: Missing If-Match header
          schema:
            $ref: "#/responses/ErrorResponse"
        500:
          description: Internal server error
          schema:
            $ref: "#/responses/ErrorResponse"
...
parameters:
  IfMatch:
    name: If-Match
    in: header
    type: string
    required: false
    description: ETag value for optimistic concurrency control
  FilePath:
    name: path
    in: query
    required: true
    type: string
    description: Absolute file path on the target system
...
responses:
  ErrorResponse:
    description: Standard error response
    schema:
      $ref: "#/definitions/Error"
  FilePropertiesResponse:
    description: File properties with metadata
    headers:
      ETag:
        type: string
        description: ETag for optimistic concurrency control
    schema:
      $ref: "#/definitions/FileProperties"
...
definitions:
  Error:
    type: object
    properties:
      code:
        type: integer
      message:
        type: string
        description: Human-readable error message
      details:
        type: string
        description: Additional error details for debugging
...
  FileProperties:
    type: object
    properties:
      mode:
        type: string
        pattern: '^[0-7]{3,4}$'
        description: File permissions in octal format (e.g., "0644")
        example: "0644"
      owner:
        type: string
      group:
        type: string
      checksum:
        type: string
        description: SHA-256 checksum of file content
...